Proper lake-indexer image and more reliable builds
