+++ integration-tests/chain-signatures/src/containers.rs
@@ -308,17 +308,11 @@ impl<'a> LakeIndexer<'a> {
             "running NEAR Lake Indexer container..."
         );
 
-        // TODO: replace with proper image
-        let image = GenericImage::new(
-            // "ghcr.io/near/near-lake-indexer",
-            // "18ef24922fd7b5b8985ea793fdf7a939e57216ba",
-            "near/near-lake-indexer",
-            "1.38",
-        )
-        .with_env_var("AWS_ACCESS_KEY_ID", "FAKE_LOCALSTACK_KEY_ID")
-        .with_env_var("AWS_SECRET_ACCESS_KEY", "FAKE_LOCALSTACK_ACCESS_KEY")
-        .with_wait_for(WaitFor::message_on_stderr("Starting Streamer"))
-        .with_exposed_port(Self::CONTAINER_RPC_PORT);
+        let image = GenericImage::new("ghcr.io/near/near-lake-indexer", "node-1.38")
+            .with_env_var("AWS_ACCESS_KEY_ID", "FAKE_LOCALSTACK_KEY_ID")
+            .with_env_var("AWS_SECRET_ACCESS_KEY", "FAKE_LOCALSTACK_ACCESS_KEY")
+            .with_wait_for(WaitFor::message_on_stderr("Starting Streamer"))
+            .with_exposed_port(Self::CONTAINER_RPC_PORT);
         let image: RunnableImage<GenericImage> = (
             image,
             vec![
